Update VIC.OLI-DELIVERABLE-3 CODE.py
This part of the code checks for missing values in the dataset and fills them in so analysis can run smoothly. Numeric missing values are replaced with the column mean, while text values are replaced with "Unknown". This ensure the dataset is complete and prevents errors in later calculations or visualizations.